Creates ASOSInfo object and moves data files
The ASOSInfo class is created, which will make it easy to find
weather stations in a given state or country; more fine scale
extraction methods can be added later such as county or even
within a given polygon.

The class takes no requried inputs, reading in the asos-stations.txt
file by default. Each line of the file is parsed in to a dictionary
with keys that match the header line of the file.

Also moves the .txt, .shp, and .shx files into the ./data directory
to keep things a little more organized.

v0.0.2