Adds function of locating radar/pixels to point
Functions for finding the closest radar to a given lat/lon and
finding radar pixels close to a lat/lon have been added.

The function for finding pixels close to a lat/lon is NOT finished.
Need to write a helper function to find corresponding radar
file; i.e., file for radar that is closest to point and closest
to datetime (probably within 10 mintues) requested by user.

v0.3.0