Adds new classes and VTEC parsing
First, minor change to WeatherRadar.py; changed the if __main__
section to allow for inputing path to shape file, if no path then
use a default
Also updated setup.py to include new dependencies

Additions include:
  vtec.py
  wwaVTEC.py
  plotUtils.py
  Some new data files for testing

vtec.py
  This file contains some dictionaries for parsing information from
the IEM NWS watch/warning shape files. Some informaiton includes
the weather phenomenon (get long name and color of plotting),
event type (watch, warning, etc.), and information about whether
event is a polygon or county warning type.
  The main thing that is important here is the PHENOM dictionary
as it sets the color for plotting the polygon in the wwaVTEC class.

wwaVTEC.py
  The wwaVTEC class is a subclass of shapely.geometry.MultiPolygon
and stores information about a single watch/warning. To initialize
class, simply pass a record from a shapefile and data will be parsed.
Currently, only useful method is plot(), which axes a matplotlib
axes object as input and plots the polygon on the axes.
  Future additions to this class could be a method to filter a list
of stations to only points that are within the warning, checking
if a datetime is within the warning datetime, determining which
NEXRAD radar site(s) contain the warning, etc.

plotUtils.py
  Contains 3 fuctions (one is already depricated).

  - baseMap  : Very basic function to set up a cartopy map with
      coastlines and U.S. state boundaries plotted. Function returns
      axes object for future use
  - plotStation : Function to plot station dot and label given
      axes to plot on, station name and coordinates
  - plotWarning : Function to plot warning box on given axes.
      This is now depricated and replaced by wwaVTEC.plot() method

v0.4.0