Updates prettyPlots for NEXRAD stations
The nexrad_station_info() function has been updated to not include
the nexrad station info text directly and now reads from the
nexrad-stations.txt file in WeatherRadarML/data/ and parses it.

The radar_range() function in prettyPlots.py has been updated to
use the NEXRAD station information instead of the ASOS station
info to create the plot.

The variable dataDir is not defined in the main __init__.py file
for the package so it is easy to reference the data directory from
other code in the package.

The setup.py file has been updated to include xarray, which is
required by one of the other dependencies; do not remember which one.

v0.6.0